Add policies and add gate for admin section.

+++ app/Http/Controllers/Web/Admin/Articles/ArticlesController.php
@@ -6,13 +6,22 @@ use App\Http\Controllers\Web\Admin\Articles\Requests\StoreArticleRequest;
 use App\Http\Controllers\Web\Admin\Articles\Requests\UpdateArticleRequest;
 use App\Models\Article;
 use App\Http\Controllers\Controller;
+use App\Policies\Abilities;
 use App\Services\Articles\ArticlesService;
 use Illuminate\Http\Request;
 
+/**
+ * Class ArticlesController
+ * @package App\Http\Controllers\Web\Admin\Articles
+ */
 class ArticlesController extends Controller
 {
     protected $articlesService;
 
+    /**
+     * ArticlesController constructor.
+     * @param ArticlesService $articlesService
+     */
     public function __construct(ArticlesService $articlesService)
     {
         $this->articlesService = $articlesService;
@@ -22,10 +31,13 @@ class ArticlesController extends Controller
      * Display a listing of the resource.
      *
      * @param Request $request
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function index(Request $request)
     {
+        $this->authorize(Abilities::VIEW_ANY, Article::class);
+
         $articleList = $this->articlesService->searchArticles($request->all());
         \View::share([
             'articleList' => $articleList
@@ -38,9 +50,12 @@ class ArticlesController extends Controller
      * Show the form for creating a new resource.
      *
      * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function create()
     {
+        $this->authorize(Abilities::CREATE, Article::class);
+
         return view('admin.articles.create');
     }
 
@@ -49,9 +64,12 @@ class ArticlesController extends Controller
      *
      * @param StoreArticleRequest $request
      * @return \Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function store(StoreArticleRequest $request)
     {
+        $this->authorize(Abilities::CREATE, Article::class);
+
         $article = $this->articlesService->storeArticle($request->getFormData());
 
         return redirect(route('admin.articles.show', $article));
@@ -62,9 +80,12 @@ class ArticlesController extends Controller
      *
      * @param Article $article
      * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function show(Article $article)
     {
+        $this->authorize(Abilities::VIEW, Article::class);
+
         return view('admin.articles.show', [
             'article' => $article
         ]);
@@ -74,10 +95,13 @@ class ArticlesController extends Controller
      * Show the form for editing the specified resource.
      *
      * @param Article $article
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function edit(Article $article)
     {
+        $this->authorize(Abilities::UPDATE, Article::class);
+
         return view('admin.articles.edit', [
             'article' => $article
         ]);
@@ -88,10 +112,13 @@ class ArticlesController extends Controller
      *
      * @param UpdateArticleRequest $request
      * @param Article $article
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function update(UpdateArticleRequest $request, Article $article)
     {
+        $this->authorize(Abilities::UPDATE, Article::class);
+
         $this->articlesService->updateArticle($article, $request->getFormData());
 
         return redirect(route('admin.articles.show', $article));
@@ -101,10 +128,13 @@ class ArticlesController extends Controller
      * Remove the specified resource from storage.
      *
      * @param Article $article
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function destroy(Article $article)
     {
+        $this->authorize(Abilities::DELETE, Article::class);
+
         $this->articlesService->deleteArticle($article);
 
         return view(
