+++ app/Http/Controllers/Web/Admin/Events/EventsController.php
@@ -6,13 +6,22 @@ use App\Http\Controllers\Controller;
 use App\Http\Controllers\Web\Admin\Events\Requests\StoreEventRequest;
 use App\Http\Controllers\Web\Admin\Events\Requests\UpdateEventRequest;
 use App\Models\Event;
+use App\Policies\Abilities;
 use App\Services\Events\EventsService;
 use Illuminate\Http\Request;
 
+/**
+ * Class EventsController
+ * @package App\Http\Controllers\Web\Admin\Events
+ */
 class EventsController extends Controller
 {
     protected $eventsService;
 
+    /**
+     * EventsController constructor.
+     * @param EventsService $eventsService
+     */
     public function __construct(EventsService $eventsService)
     {
         $this->eventsService = $eventsService;
@@ -22,10 +31,13 @@ class EventsController extends Controller
      * Display a listing of the resource.
      *
      * @param Request $request
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function index(Request $request)
     {
+        $this->authorize(Abilities::VIEW_ANY, Event::class);
+
         $eventList = $this->eventsService->searchEvents($request->all());
         \View::share([
             'eventList' => $eventList
@@ -37,21 +49,26 @@ class EventsController extends Controller
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function create()
     {
+        $this->authorize(Abilities::CREATE, Event::class);
+
         return view('admin.events.create');
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param \Illuminate\Http\Request $request
-     * @return \Illuminate\Http\Response
+     * @param StoreEventRequest $request
+     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function store(StoreEventRequest $request)
     {
+        $this->authorize(Abilities::CREATE, Event::class);
         $event = $this->eventsService->storeEvent($request->getFormData());
 
         return redirect(route('admin.events.show', $event));
@@ -62,9 +79,12 @@ class EventsController extends Controller
      *
      * @param Event $event
      * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function show(Event $event)
     {
+        $this->authorize(Abilities::VIEW, Event::class);
+
         return view('admin.events.show', [
             'event' => $event
         ]);
@@ -74,10 +94,13 @@ class EventsController extends Controller
      * Show the form for editing the specified resource.
      *
      * @param Event $event
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function edit(Event $event)
     {
+        $this->authorize(Abilities::UPDATE, Event::class);
+
         return view('admin.events.edit', [
             'event' => $event
         ]);
@@ -88,10 +111,12 @@ class EventsController extends Controller
      *
      * @param UpdateEventRequest $request
      * @param Event $event
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function update(UpdateEventRequest $request, Event $event)
     {
+        $this->authorize(Abilities::UPDATE, Event::class);
         $this->eventsService->updateEvent($event, $request->getFormData());
 
         return redirect(route('admin.events.show', $event));
@@ -101,10 +126,12 @@ class EventsController extends Controller
      * Remove the specified resource from storage.
      *
      * @param Event $event
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function destroy(Event $event)
     {
+        $this->authorize(Abilities::DELETE, Event::class);
         $this->eventsService->deleteEvent($event);
 
         return view(
